% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.errs.R
\name{barplot_errs}
\alias{barplot_errs}
\title{Setting for computing scales of a few tests}
\usage{
barplot_errs(X = NULL, grps = NULL, title = "Means", col.palette = NULL,
  na.rm = TRUE, filename = NULL, sort.by = "order=order(colMeans(X))",
  means.cis = NULL)
}
\arguments{
\item{X}{NULL}

\item{grps}{NULL}

\item{title}{"Means"}

\item{col.palette}{NULL}

\item{filename}{NULL}

\item{sort.by}{"order=order(colMeans(X))"}

\item{means.cis}{NULL}

\item{na.r}{TRUE}
}
\description{
barre d'errore
}

