% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_realized_fwer_and_power.R
\name{compute_realized_fwer_and_power}
\alias{compute_realized_fwer_and_power}
\title{compute_realized_fwer_and_power}
\usage{
compute_realized_fwer_and_power(ps, h1 = rep(FALSE, length(ps)),
  alpha = 0.05)
}
\description{
compute_realized_fwer_and_power
}

