% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillNA.R
\name{fillNA}
\alias{fillNA}
\title{fill NAs of a column using all other columns (as covariates in a lm)}
\usage{
fillNA(data, name.fill)
}
\arguments{
\item{data}{a datsa.frame}

\item{name.fill}{the column name to fill the NAs using all other columns}
}

