% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_range}
\alias{in_range}
\title{Is/Are the value(s) in x within the reange defined by lims?}
\usage{
in_range(x, lims, equal_included = c(TRUE, FALSE))
}
\description{
Is/Are the value(s) in x within the reange defined by lims?
}
