% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxt_light.R
\name{maxT.light}
\alias{maxT.light}
\title{(stepdown) Westfall and Young max-t procedure}
\usage{
maxT.light(
  permT,
  alphas = c(0.001, 0.01, 0.05, 0.1),
  weights = NULL,
  m = ncol(permT)
)
}
\description{
(stepdown) Westfall and Young max-t procedure
}
