% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_any_test_DA INCLUDERE.R
\name{pairwise.any.test}
\alias{pairwise.any.test}
\title{pairwise.any.test with post hoc correction. Multivariate output is allowed. The results are stored in the third dimension of the array.}
\usage{
pairwise.any.test(test_fun, groups, data, p.adjust.methods = "Shaffer")
}
\arguments{
\item{test_fun}{a function that has data as input and a p-value as output. it is computed for each pair}

\item{groups}{a vector of group ids}

\item{data}{a data.frame or matrix to be use by `test_fun`}

\item{p.adjust.methods}{any method used by p.adjust + "Shaffer" + "closed" testing (in this case `test_fun` is used in every element of the closure set. NOT implemented, yet!)}
}
\description{
pairwise.any.test with post hoc correction. Multivariate output is allowed. The results are stored in the third dimension of the array.
}
