% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_chol.R
\name{solve_chol}
\alias{solve_chol}
\title{solve_chol}
\usage{
solve_chol(M)
}
\arguments{
\item{M}{symmetric matrix to invert}
}
\description{
matrix inversion using cholesky decomp. with backward substitution
}
