% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsExplore.r
\name{stats}
\alias{stats}
\alias{tableFrPrc}
\title{descriptive stats functions}
\usage{
stats(
  x,
  stat = c("mean", "sd", "valids", "not.valids", "quantile", "CI", "mse"),
  as.vector = FALSE
)
}
\arguments{
\item{x}{vettore}

\item{stat}{tipo di statistica test}

\item{as.vector}{logico}

\item{title}{NULL}

\item{rounding}{1 di default}
}
\description{
descriptive stats functions
}
