% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsExplore.r
\name{strata.meanPmSd}
\alias{strata.meanPmSd}
\title{strata.meanPmSd}
\usage{
strata.meanPmSd(Y, strata, round.digits = 1, na.rm = FALSE, test = FALSE,
  mse = FALSE, as.table = FALSE, pm = TRUE, addTotal = FALSE,
  medianAndRange = FALSE, as.character = TRUE)
}
\arguments{
\item{Y}{matice}

\item{strata}{vettore o NULL}

\item{round.digits}{1}

\item{na.rm}{FALSE}

\item{test}{FALSE}

\item{mse}{FALSE}

\item{as.table}{FALSE}

\item{pm}{TRUE}

\item{addTotal}{FALSE}

\item{medianAndRange}{FALSE}

\item{as.character}{TRUE}
}
\description{
strata.meanPmSd
}

