% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_light.R
\name{t.test.light}
\alias{t.test.light}
\title{t.test.light}
\usage{
\method{t}{test.light}(Y, tail = 1)
}
\arguments{
\item{Y}{n X p matrix or data.frame}

\item{tail}{+1,-1 or 0}
}
\description{
t.test.light
}
