% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webplot.R
\name{webplot}
\alias{webplot}
\alias{webplot.multi}
\title{webplot}
\usage{
webplot(
  data,
  data.row = NULL,
  y.cols = NULL,
  main = NULL,
  add = F,
  col = "red",
  scale = T,
  lwd = 3,
  lty = 1,
  cex = 1,
  ...
)
}
\arguments{
\item{data}{to be done}

\item{data.row}{to be done}

\item{y.cols}{to be done}

\item{main}{to be done}

\item{add}{to be done}

\item{col}{as in \code{plot()}}

\item{scale}{if \code{TRUE}: \code{data = apply(data, 2, function(x) ((x -
min(x))/(max(x)-min(x))+.1)/1.1)}}

\item{lwd}{as in \code{plot()}}

\item{lty}{as in \code{plot()}}

\item{cex}{as in \code{plot()}}

\item{\dots}{to be done}
}
\description{
webplot
}
\examples{
data(pal.unipd)

X <- as.data.frame(matrix(rpois(30,100),3,10))
webplot(X[1,])
webplot.multi(X,col=pal.unipd)

}
