% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.errs.R
\name{barplot_errs}
\alias{barplot_errs}
\title{Setting for computing scales of a few tests}
\usage{
barplot_errs(X = NULL, grps = NULL, title = "Means", col.palette = NULL,
  na.rm = TRUE, filename = NULL,
  sort.by = "order.vars=order(colMeans(X))", means.cis = NULL,
  ylab = "means", xlab = "", grp.names = "grps")
}
\arguments{
\item{X}{NULL}

\item{grps}{NULL}

\item{title}{"Means"}

\item{col.palette}{NULL}

\item{na.rm}{TRUE}

\item{filename}{NULL}

\item{sort.by}{"order.vars=order(colMeans(X))"}

\item{means.cis}{NULL}
}
\description{
barre d'errore
}
\examples{
\dontrun{
plt=barplot_errs(dat,col.palette=wes.palette(2, "Darjeeling"),grps=dati$pz.,title=title,sort.by=sort.by)     
ps.cor=apply(dat,2,function(x)cor.test(dati$pz,x)$p.value)
signif =names(which(p.adjust(ps.cor,"holm")<=.05))
plt=plt+ annotate("text", x = signif, y = -.12, label = "*",color="darkred",size=15)
plt
}
}

