% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{cc}
\usage{
cc(X, Y)
}
\description{
Same as \code{cc()} of package \code{CCA}, but it allows for X and Y to be rank deficient.
See \code{\link[CCA]{cc}} for a proper documentation.
}
